requirements.txt
streamlit>=1.28.0
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0